requirements.txt
Flask
requests
python-dotenv
flasgger
